Location permission to Always change no 1
When user selects "When in use" appropriate prompt is spawned and user is immediately transfered to settings where he is
"forced" to choose "Always". Starting with iOS 13, Apple introduced a new location access prompt that only allows users to
grant location access permissions in the "While Using the App" mode initially. To request location access in the "Always" mode,
you need to follow the same approach as before, but once the user grants the "While Using the App" permission, you can then
prompt the user to change the location access mode to "Always" manually.